+++ input_main.cpp
@@ -26,17 +26,17 @@ int main(int argc, char *argv[])
     
     StateProb2 test(4);
     cout<<test;
-    double pr=0.5;
+    long double pr=0.5;
     int flowNum=4,ruleNum=3;
     int mSize=1;
-    double msizerate=0.7;
+    long double msizerate=0.7;
     StateType initialStateNum=0;
-    double interval=2;
-   	double limit=0.001;
-    double delta=0.001;
+    long double interval=2;
+   	long double limit=0.001;
+    long double delta=0.001;
     int target=1;
     int qnum=1;
-    double TTLmax=1;
+    long double TTLmax=1;
     int runtimes=10;
     string label="";
     if(argc>1)
@@ -56,7 +56,7 @@ int main(int argc, char *argv[])
     FlowRuleTable table(flowNum,ruleNum);//flowNUm,ruleNUm
     floatCounter flowPara(flowNum);
     floatCounter TTL(ruleNum);
-    double unit=0;//=unitComputation(&flowPara, delta, limit, &TTL);
+    long double unit=0;//=unitComputation(&flowPara, delta, limit, &TTL);
      Automatic a(label,&flowPara,&table,&TTL,mSize,initialStateNum,interval,unit,delta);
     
     int interestflow=a.generate(flowNum, ruleNum, msizerate, TTLmax,interval,runtimes);
